// src/app/(admin)/admin/media-library/types.ts
export type MediaResource = {
  public_id: string;
  secure_url: string;
  format: string;
  width: number;
  height: number;
  bytes: number;
  created_at: string;
  resource_type: string;
};

export type MediaFolder = {
  name: string;
  path: string;
};

export type MediaLibraryData = {
  resources: MediaResource[];
  folders: MediaFolder[];
  currentFolder: string;
};